Fix a terrible  BUG!
In last submit, the program will launch a lot of word or excel processes. So, it will eat up your memory as your target files increase.
In this submit, I fix this problem. The reason is the Quit() method in win32com.client.Dispatch('Word.Application') doesn't close the word process when a sub job is done, so Excel. del method is added to really delete the process.